Move SalvaContext to salva_context.rs
