+++ src/plugin/mod.rs
@@ -1,8 +1,10 @@
 pub use crate::fluid::AppendNonPressureForces;
 pub use crate::fluid::RemoveNonPressureForcesAt;
 pub use self::plugin::{
-    SalvaContext, SalvaPhysicsPlugin, SalvaSimulationSet
+    SalvaPhysicsPlugin, SalvaSimulationSet
 };
+pub use salva_context::*;
 
 pub mod systems;
 mod plugin;
+mod salva_context;
